Refactor states.rb
Separate Item class from lrama/states.rb
